Implement Command pattern
The Command pattern is a behavioral design pattern in which an object is
used to encapsulate all information needed to perform an action. The
Command pattern is one of the bases of the Service Layer pattern.

Provide an interface to structure and implement a command object. It
imposes the implementation of a perform method and provides a delegation
of the execution logic of the command.